Safari-611.1.10/LayoutTests/accessibility/ios-simulator/tab-role-expected.txt

tab1

This test makes sure that role='tab' shows up as a distinct element from button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tab.description is 'AXLabel: tab1'
PASS button.traits != tab.traits is true
PASS tab.traits != 0 is true
PASS tab.isSelected is true
PASS successfullyParsed is true

TEST COMPLETE
